Create README.md
The goal is to fill the maximum number of contours (German: Umrisse) by falling pieces of different shapes -- Tetris stones (German: Steine), or "polyomino". These stones are randomly generated and have certain sizes (k*k).

Each contour is a combination of two polyform stones. There are also parking spaces (German: Parkplatz),where stones can be stored ("parked") if there is no time to put them in the contour or a stone in the contour is occupied.

With a user interface the player can choose the number of falling pieces, their size (k), speed of the game, the number of contours and parking spaces, colours, and acceleration (yes or now).With the mouse, the player can select one of the falling stones, manipulate it (rotate, reflect) and place it on the contour or parking space, which are also selected using the mouse or keyboard. As soon as at least one falling stone touches the bottom edge of the playing field, the game is over. Detailed rules of the game can be found in the instructions for the game (Spielanleitung).